Move Rebol2 UPARSE combinators into Redbol
Having the compatibility parsers in the main build is not
necessary.  However, there needs to be some attention given
to having a Redbol test suite.  (Currently the only Redbol
tests are a couple of historical codebases run along with
their modern counterparts on GitHub Actions.)

This moves the UPARSE2 tests into the %tests/ directory, but
really it should be running tests from Red.