Move REDBOL-APPLY from core tests to Redbol
Development on the experimental REDBOL-APPLY was being done in
the core respository pending the development of a more modern
native form of APPLY, just to show the FRAME! mechanics working.

A snapshot of that had been copied into the Redbol repository,
but it had gotten out of sync...so the one in Redbol was no
longer working.

This makes Redbol the new "official" home of the construct, and
moves its tests as well.  It is put in a separate module with
the FUNC and FUNCTION emulation code.