Use @none instead of BLANK! for none in Redbol
The new rule in Ren-C is that BLANK! evaluates to null, and it
is truthy in its reified form.

So a new answer is needed for doing a reified non-evaluating
falsey thing in Redbol.  This uses the THE-WORD! of @none,
and then supplies a /PREDICATE for truthiness to CASE and such
that considers it falsey.