Temporary change to EVALUATE
When multi-returns were added as a feature they were used
many places, including functions that piped evaluations.  This
became a bit more awkward when multi-returns shifted to an
implementation based on parameter packs, because the intent
of a multi-return of the piped thing became conflated with
the multi-return intent of the piper.

This is being rethought, but for the moment some routines
are trying going back to passing in a variable as an
argument.  A better answer for EVALUATE is likely building
some kind of state (an evaluator object) to use--which is
needed for other reasons.